move related structs to an appropriated module
